Refactored to use a mixin approach (almost) for leaf cells
