+++ lib/src/synthesizers/schematic/schematic.dart
@@ -4,9 +4,9 @@
 export 'constant_handler.dart';
 export 'module_utils.dart';
 export 'passthrough_handler.dart';
+export 'schematic_gates.dart';
 export 'schematic_mixins.dart';
-export 'schematic_primitives.dart';
 export 'schematic_synthesis_result.dart';
 export 'schematic_synthesizer.dart';
-export 'sequential_handler.dart';
+export 'sequential_schematic.dart';
 export 'yosys/yosys_loader_helper.dart';
